Point inside of a box
I am currently starting the process of creating an interaction system which requires knowledge of whether a point is inside of a box (since there is no objects). This is the start of the system where I have created an algorithm for determining if a point is inside of a box based on research. I have no clue whether it works yet though as there is still some more set up to do with the procedure that needs to be done.