Point in box Cont
Went through and deleted everything deciding to try and use parts of the collision system I previously created because the other one was proving to be more difficult than intended. Plus this way we can keep our code base smaller with using the same functions. This may even require the use of function pointers in the future, yipee!