Refactoring collision
In an effort to reuse code efficiently, I've started to refactor some of the collision system's code. Currently, with the changes I've made so far, the system works perfectly with regular collision detection and blocking response. The problem is with the is point in box which isn't reporting as true despite the code being seemingly similar. This is a tomorrow problem.