Collision has been refactored for now
Turns out the problem was the < in an area of the code. Collision code has been refactored for now, more refactoring will come as needed but all I really wanted to do was refactor getting our T values as that was the code that was mainly used and will be used again for our ray casting procedures in a little.