Interaction system in progress
Created the interaction system so that code fires off when 'e' is pressed. Unsure of if it works or not that is to be determined. I'll probably make a system that allows me to visualize the player's interact trace but thats gonna require linked lists so I'm submitting now before I start making that exciting venture