Problems with interaction
As suspected, the interaction system isn't complete as I had hoped it would be. Something is wrong with the intersection point as it ends up in a weird place for certain parts of the cube. Could be the result of an improper calculation of the intersectionpoint or the calculation of the triangle. Other than this realization, the Query areas for the player have started to have been set up, there are some small things still left like actually using it and some probable bugs but I want to figure out this bug with the interaction system before I continue work on this