Button construction complete?
Well that was deceptively too easy. The buttons load properly and exist in a grid as wanted, they aren't the right color but I'm guessing that has to do with my inability to UV. I also need to free the buttons on game close, but other than that their existing functionality is seemingly working.