Mechanic in progress, re looking at interaction
Interaction system has some bugs still, not always getting the intersection point quite correct, added in checking for verts and edges which made it fail a lot less but it still fails nonetheless. Mechanic is in progress with the interaction sometimes working. Might continue to work on the mechanic and come back to the troubling interaction in the future since I currently have no solutions for the problem and the interaction will work some of the time.