Interaction and puzzle combination
Continued work on the mechanic. Bridging the game between the interaction system and the puzzle mechanic so that we can interact with puzzles and go into a puzzle mode that allows you to move the highlight around with the arrow keys. Currently we crash whenever we interact, working to figure that out.