File reading for puzzle solutions
Currently working on a system that will read csv files with the locations of the buttons that are the ones to get to the solution. This will make it easier to prototype different ideas by making it really simple to create a csv file of the rows and columns with the location of the correct buttons, and then load this file in with the call of a function and the solution fills itself in.