The start of things with solutions to buttons
Alrighty, so now the solution can be loaded into via CSV file, there might be some errors there because I have not checked it against whether an unmatching amount of columns and rows will work. Also, started to work on getting the solution of the puzzle but there is an error in the fact that I can put more than what is the answer and yet still get told I got the solution.