Mechanic refinement
The button puzzle was created on the wrong axis. To solve this, I turned the button in the obj file, and moved some things around in the creation of the buttons. Now things should work as intended, rotating the master should be correct now that they are oriented on the correct axis. Also another lesson, always initialize struct values to 0 or null or something besides the garbage values that they will hold.