I'm calling it here
I'm not going to spend more time trying to figure out what's wrong with the buttons, I need to move on. At some point, I will go through the entire project and determine where there could be memory leaks and fix it. But for now I must continue building more of the project.