Decided to continue another way
Using csv files was proving to cause more issues than good. It'll probably be something I come back to when I have the time to, but for the time being, I decided to just go to manually inputting the locations rather than importing csv file locations for the solutions of the puzzles. Currently your puzzle_03 locations are incorrect and it's something you need to fix.