Reworked puzzle rule, working on working it in
After reworking the original rule for the first puzzle, I am now starting to work on creating for that rule. I also removed the reading of csv files for solutions as it was causing too many issues and now you are required to input them manually... I may have already said that, but this submit was also just a further refinement on that idea as well.