Button Textures
I swapped the button textures to now exist in a struct so that they can be swapped via reference rather than unloading and loading the textures constantly after every button press which probably would have stacked up at some point. Plus it makes it easier to implement the custom textures for the symbols on the buttons.