Button Powered off functionality
Started on the functionality of the button powered off state. Currently there is a state in which you can implement in which the buttons are textured in a powered off state and the entire puzzle cannot be interacted with until it is powered on. What I need to do next is create the ability to power the puzzle on when another puzzle is completed.