New texture system implemented
Textures are now only loaded in once on the beginning of the game and are located in a file in one place. Accessing these textures is done by accessing the global array from main.c in which you can grab the texture needed by grabbing the specific index of the texture as written in the file where the array of file locations exists.