Button collateral selection
Creating the process of allowing other toggle buttons that are in rows of other toggle buttons to be selected when one of them is. Currently submitting because I'm about to mess about with some functionality about how the puzzle originally worked to see if I like this solution better than what is currently in place.