Player lerping
Currently working on a system that lerps the player in front of the puzzle when they interact with it. This will become more evidently useful later in the project when I flesh out the ideas I have had planned for a while. Also an excuse to not work on gameb though I should be.