Progress on gameb
Continuing to work on gameb puzzles, more comfortable with the way the system works now that it's been iterated a couple of times - doesn't mean there won't be any more iteration. After I'm comfortable with the amount of puzzles in game b, I'll be going back to gamea to fix its intro to the puzzles and after that start on multiple windows!