We have 2 games!
Gamea creates a child process for gameb so that we can utilize shared memory. They do share memory atm but it's just an integer and while I haven't found a way to tell if the memory has truly be shared, according to the example I used to create this project, it should be working fine. Now it's time to devise a plan to design a system that incorporates the shared memory.