Shared Puzzles cont
So far, the puzzles work when you are interacting with them and the solutions are updated and all whenever there is a shared puzzle. However if you are walking around in one area and try to connect to a puzzle that already has some submitted buttons, the puzzle won't update. Check TODO to see where you are in the progress if you can't remember where you left off when you come back.